automation/data-generator/requirements.txt
httpx
asyncio
python-jose[cryptography]